Panel mounts for 1/8" panels with ~1/16" foam tape
These files were specifically modified to use the same size hardware as the stock mounts. The front panel latch bodies also do not require any changes to the rest of the lock parts.

The clips with 'pins' designation, have pins on either side of the bolt designed to bottom out in misumi extrusions so the clips can be properly tightened without over-compressing the foam tape.